Move `FoldersViewModel` to `: feature:navigation:drawer`
